+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Try to keep listed changes to a concise bulleted list of simple explanations of 
 
 ### Changed
 - The serverconfig was moved from internal to pkg to make it available to external users of this package. [#2382](https://github.com/openfga/openfga/pull/2382)
+- Panics in goroutines in check are converted to errors.
 
 ### Fixed
 - Add limit to goroutine concurrency when processing iterator [#2386](https://github.com/openfga/openfga/pull/2386)